Add Policies & Permissions

+++ app/Http/Controllers/Api/V1/TicketController.php
@@ -10,11 +10,16 @@ use App\Http\Requests\Api\V1\StoreTicketRequest;
 use App\Http\Requests\Api\V1\UpdateTicketRequest;
 use App\Http\Resources\V1\TicketResource;
 use App\Models\User;
+use App\Policies\V1\TicketPolicy;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class TicketController extends ApiController
 {
+
+    protected $policyClass = TicketPolicy::class;
     /**
      * Display a listing of the resource.
      */
@@ -30,14 +35,14 @@ class TicketController extends ApiController
     public function store(StoreTicketRequest $request)
     {
         try {
-            $user = User::findOrFail($request->input('data.relationships.author.data.id'));
-        } catch (ModelNotFoundException $exception) {
-            return $this->ok('User not found', [
-                'error' => 'The provided user id does not exists'
-            ]);
-        }
 
-        return new TicketResource(Ticket::create($request->mappedAttribtes()));
+            Gate::authorize('store', Ticket::class); 
+
+            return new TicketResource(Ticket::create($request->mappedAttributes()));
+
+        } catch (AuthorizationException $ex) {
+            return $this->error('You are not authorized to update that resource.', 401);
+        }
     }
 
     /**
@@ -66,6 +71,9 @@ class TicketController extends ApiController
 
             $ticket = Ticket::findOrFail($ticket_id);
 
+            //policy
+            Gate::authorize('update', $ticket);  
+
             $ticket->update($request->mappedAttributes());
 
             return new TicketResource($ticket);
@@ -73,6 +81,8 @@ class TicketController extends ApiController
             return $this->ok('User not found', [
                 'error' => 'The provided user id does not exists'
             ]);
+        } catch (AuthorizationException $ex) {
+            return $this->error('You are not authorized to update that resource.', 401);
         }
     }
 
@@ -84,6 +94,8 @@ class TicketController extends ApiController
         try {
 
             $ticket = Ticket::findOrFail($ticket_id);
+
+            Gate::authorize('replace', $ticket); 
 
             $ticket->update($request->mappedAttributes());
 
@@ -102,6 +114,9 @@ class TicketController extends ApiController
     {
         try {
             $ticket = Ticket::findOrFail($ticket_id);
+
+            Gate::authorize('update', $ticket);  
+
             $ticket->delete();
 
             return $this->ok('Ticket successfully deleted');
